refactor: split utils into separate files
